+++ app/components/validated-input/component.js
@@ -1,0 +1,2 @@
+
+export { default } from 'ember-osf/components/validated-input/component';